Consolidate backend interfaces
