+++ packages/lms-client/package.json
@@ -11,12 +11,10 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@lmstudio/lms-common": "^0.6.1",
-    "@lmstudio/lms-diagnostics-backend-interface": "^0.1.1",
     "@lmstudio/lms-external-backend-interfaces": "^0.1.0",
     "@lmstudio/lms-isomorphic": "^0.3.2",
     "@lmstudio/lms-kv-config": "^0.1.0",
     "@lmstudio/lms-shared-types": "^0.5.1",
-    "@lmstudio/lms-system-backend-interface": "^0.3.1",
     "boxen": "^5.1.2",
     "chalk": "^4.1.2",
     "process": "^0.11.10",
