+++ packages/lms-client/src/LMStudioClient.ts
@@ -17,25 +17,21 @@ import {
 import { getHostedEnv, type ClientPort } from "@lmstudio/lms-communication-client";
 import {
   createDiagnosticsBackendInterface,
-  type DiagnosticsPort,
-} from "@lmstudio/lms-diagnostics-backend-interface";
-import {
   createEmbeddingBackendInterface,
   createLlmBackendInterface,
   createRetrievalBackendInterface,
+  createSystemBackendInterface,
+  type DiagnosticsPort,
   type EmbeddingPort,
   type LLMPort,
   type RetrievalPort,
+  type SystemPort,
 } from "@lmstudio/lms-external-backend-interfaces";
 import {
   createFilesBackendInterface,
   type FilesPort,
 } from "@lmstudio/lms-external-backend-interfaces/dist/filesBackendInterface";
 import { generateRandomBase64 } from "@lmstudio/lms-isomorphic";
-import {
-  createSystemBackendInterface,
-  type SystemPort,
-} from "@lmstudio/lms-system-backend-interface";
 import chalk from "chalk";
 import process from "process";
 import { z } from "zod";
