--- packages/lms-diagnostics-backend-interface/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@lmstudio/lms-diagnostics-backend-interface",
-  "version": "0.1.1",
-  "description": "",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "build": "tsc"
-  },
-  "author": "",
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@lmstudio/lms-communication": "^0.5.1",
-    "@lmstudio/lms-shared-types": "^0.5.1",
-    "zod": "^3.22.4"
-  },
-  "gitHead": "96f419904258c472237e2a50496711c7552ed8af"
-}
