+++ tsconfig.build.json
@@ -11,9 +11,6 @@
     { "path": "packages/lms-communication-server" },
     { "path": "packages/lms-communication-mock" },
     { "path": "packages/lms-external-backend-interfaces" },
-    { "path": "packages/lms-platform-backend-interface" },
-    { "path": "packages/lms-system-backend-interface" },
-    { "path": "packages/lms-diagnostics-backend-interface" },
     { "path": "packages/lms-client" },
     { "path": "packages/lms-lmstudio" },
     { "path": "packages/template" }
